+++ CHANGELOG.md
@@ -10,6 +10,7 @@ You can find and compare releases at the GitHub release page.
 
 ### Added
 - Different TTL configurations for each guard
+- lcobucci/jwt: add array support for `aud` claim
 
 ## [2.0.0] 2022-09-08
 - No changes to 2.0.0-RC1